move couch processing code to the couch backend
