+++ corehq/apps/commtrack/processing.py
@@ -7,7 +7,7 @@ from dimagi.utils.decorators.log_exception import log_exception
 
 from casexml.apps.case.const import CASE_ACTION_COMMTRACK
 from casexml.apps.case.exceptions import IllegalCaseId
-from casexml.apps.stock.models import StockTransaction, StockReport
+from casexml.apps.stock.models import StockTransaction
 from corehq.form_processor.casedb_base import AbstractCaseDbCache
 from corehq.form_processor.interfaces.processor import FormProcessorInterface
 from casexml.apps.stock import const as stockconst
@@ -50,37 +50,6 @@ class StockProcessingResult(object):
                 processor.create_models_for_stock_report_helper(stock_report_helper)
 
 
-@transaction.atomic
-def delete_models_for_stock_report(domain, stock_report_helper):
-    """
-    Delete all stock reports and stock transaction models associated with the helper from the database.
-    """
-    assert stock_report_helper.domain == domain
-    StockReport.objects.filter(domain=domain, form_id=stock_report_helper.form_id).delete()
-
-
-@transaction.atomic
-def create_models_for_stock_report(domain, stock_report_helper):
-    """
-    Save stock report and stock transaction models to the database.
-    """
-    assert stock_report_helper.domain == domain
-    if stock_report_helper.tag not in stockconst.VALID_REPORT_TYPES:
-        return
-    report = _create_model_for_stock_report(domain, stock_report_helper)
-    for transaction_helper in stock_report_helper.transactions:
-        _create_model_for_stock_transaction(report, transaction_helper)
-
-
-def _create_model_for_stock_report(domain, stock_report_helper):
-    return StockReport.objects.create(
-        form_id=stock_report_helper.form_id,
-        date=stock_report_helper.timestamp,
-        type=stock_report_helper.tag,
-        domain=domain,
-        server_date=stock_report_helper.server_date,
-    )
-
 LedgerValues = namedtuple('LedgerValues', ['balance', 'delta'])
 
 
@@ -111,38 +80,6 @@ def compute_ledger_values(lazy_original_balance, report_type, quantity):
     else:
         raise ValueError()
     return LedgerValues(new_balance, new_delta)
-
-
-def _create_model_for_stock_transaction(report, transaction_helper):
-    assert report.type in stockconst.VALID_REPORT_TYPES
-    txn = StockTransaction(
-        report=report,
-        case_id=transaction_helper.case_id,
-        section_id=transaction_helper.section_id,
-        product_id=transaction_helper.product_id,
-        type=transaction_helper.action,
-        subtype=transaction_helper.subaction,
-    )
-
-    def lazy_original_balance():
-        previous_transaction = txn.get_previous_transaction()
-        if previous_transaction:
-            return previous_transaction.stock_on_hand
-        else:
-            return None
-
-    new_ledger_values = compute_ledger_values(
-        lazy_original_balance, report.type,
-        transaction_helper.relative_quantity)
-
-    txn.stock_on_hand = new_ledger_values.balance
-    txn.quantity = new_ledger_values.delta
-
-    if report.domain:
-        # set this as a shortcut for post save signal receivers
-        txn.domain = report.domain
-    txn.save()
-    return txn
 
 
 @log_exception()
